Authentication, Users and Permissions Part 2 Complete

+++ controllers/courses.js
@@ -48,6 +48,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access  Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId;
+  req.body.user = req.user.id;
 
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
@@ -55,6 +56,16 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
     return next(
       new ErrorResponse(`No bootcamp with id of ${req.params.bootcampId}`),
       404
+    );
+  }
+
+  // make suer tbe course owner is the logger in user
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to  ${bootcamp._id} bootcamp`,
+        401
+      )
     );
   }
 
@@ -77,6 +88,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`No course with the id of ${req.params.id}`),
       404
     );
+
+  // make suer tbe course owner is the logger in user
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
 
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -102,6 +123,15 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
       404
     );
 
+  // make suer tbe course owner is the logger in user
+  if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    );
+  }
   await course.deleteOne();
 
   res.status(200).json({
